feat(interceptor): add user id and update user with shop id
BREAKING CHANGE: on shop creation this interceptor will add user id in
to the req.body and after shop creation the shop id will be added to the
user and user will be updated.

BUGS: previously guards in user controller was ordered incorrectly. in
auth service option for saving shop_id in jwt token is now added.